Generate a raster of mean fetch lengths from a binary land-water raster.

Given a binary land and water raster, this function will calculate the mean distance a water cell is away from land. The fetch distance is the mean distance between the given water cell and the next land cell in 8 directions (horizontally, vertically, and diagonally). Returns a raster of the same CRS, resolution, and extent as the input land water raster.

Usage

get_fetch(r = NULL, max_dist = 1e+05, ncores = 2, verbose = TRUE)

Arguments

r

SpatRaster, RasterLayer or file path (character) to binary land type raster TIF. Land and water values only.

max_dist

numeric maximum distance (meters) to calculate fetch values out to. Default is 100000m (100km)

ncores

numeric indicating how many cores to use during parallel processing. Default is 2.

verbose

logical, whether messages should be printed. Default is FALSE, no messages print.

Value

SpatRaster with the mean fetch distance from land calculated for every water pixel.